MA-24331 [Android-16.0_1.4.0] WCS 26 Q1 release patch integrate
Reason: Integrate WCS 26Q1 RC2 release patch, WiFi driver part

Driver commit for release MM6X18542.p8_V1 2026-01-29_23:48:25

Firmware version for commit::
W9098-17.92.1.p149.84
IW416-16.92.21.p155.1
W8987-16.92.21.p155.1
IW612-18.99.8.p3
IW610-18.99.8.p3
AW693-18.99.2.p145.172

+++ mlinux/moal_proc.c
@@ -3,7 +3,7 @@
  * @brief This file contains functions for proc file.
  *
  *
- * Copyright 2008-2022, 2025 NXP
+ * Copyright 2008-2022, 2025-2026 NXP
  *
  * This software file (the File) is distributed by NXP
  * under the terms of the GNU General Public License Version 2, June 1991
@@ -1364,7 +1364,7 @@ static int woal_ssu_dump_read(struct seq_file *sfp, void *data)
 	if (sfp->size < ((handle->ssu_dump_len * 9) / 4)) {
 		PRINTM(MCMND,
 		       "ssu dump size too big, size=%lu, ssu_dump_len=%lu\n",
-		       sfp->size,
+		       (unsigned long)sfp->size,
 		       (unsigned long)((handle->ssu_dump_len * 9) / 4));
 		sfp->count = sfp->size;
 		ret = 0;
